+++ src/Controller/CreateProjectController.php
@@ -2,7 +2,7 @@
 
 namespace App\Controller;
 
-use App\Form\CreateProjectForm;
+use App\Form\CreateProjectType;
 use App\Service\ApiServiceInterface;
 use Symfony\Bundle\FrameworkBundle\Controller\AbstractController;
 use Symfony\Component\HttpFoundation\Request;
@@ -15,6 +15,8 @@ use Symfony\Component\Routing\Annotation\Route;
 #[Route('/admin/create-project')]
 class CreateProjectController extends AbstractController
 {
+    private const SESSION_KEY = 'create_project_form_data';
+
     public function __construct(
         private readonly ApiServiceInterface $apiService
     ) {
@@ -22,11 +24,13 @@ class CreateProjectController extends AbstractController
 
     /**
      * Create a project and all related entities.
+     *
+     * TODO: Refactor code to move Jira specifics into JiraApiService.
      */
     #[Route('/new', name: 'create_project_form')]
     public function createProject(Request $request): Response
     {
-        $form = $this->createForm(CreateProjectForm::class);
+        $form = $this->createForm(CreateProjectType::class);
         $form->handleRequest($request);
 
         // Set form data.
@@ -40,8 +44,6 @@ class CreateProjectController extends AbstractController
 
         // Handle form submission.
         if ($form->isSubmitted() && $form->isValid()) {
-            // Do stuff on submission.
-
             // Set selected team config.
             foreach ($formData['projectCategories'] as $team) {
                 if ($team->id === $formData['form']['team']) {
@@ -104,12 +106,12 @@ class CreateProjectController extends AbstractController
             $this->apiService->addProjectToTimeTrackerAccount($project, $account);
 
             // Create project board
-            if (!empty($formData['selectedTeamConfig']['board_template'])) {
+            if (!empty($formData['selectedTeamConfig']) && !empty($formData['selectedTeamConfig']['board_template'])) {
                 $this->apiService->createProjectBoard($formData['selectedTeamConfig']['board_template']['type'], $project);
             }
 
-            // Go to form submitted page.
-            $_SESSION['form_data'] = $formData;
+            $session = $request->getSession();
+            $session->set(self::SESSION_KEY, $formData);
 
             return $this->redirectToRoute('create_project_submitted');
         }
@@ -131,14 +133,39 @@ class CreateProjectController extends AbstractController
     /**
      * Receipt page displayed when a project was created.
      *
+     * TODO: Refactor code to move Jira specifics into JiraApiService.
+     *
      * @Route("/submitted", name="create_project_submitted")
      */
     public function submitted(Request $request): Response
     {
+        // Get session data, and clean session entry.
+        $session = $request->getSession();
+        $formData = $session->get(self::SESSION_KEY);
+        $session->remove(self::SESSION_KEY);
+
+        $endpoints = $this->apiService->getEndpoints();
+        $url = $endpoints['base'];
+        $url = is_string($url) ? $url : '';
+        $teamId = $formData['selectedTeamConfig']['tempo_team_id'] ?? '';
+
         return $this->render(
             'create_project/submitted.html.twig',
             [
-                'form_data' => $_SESSION['form_data'],
+                'url' => "$url/secure/Tempo.jspa#/teams/team/$teamId/" ?? null,
+                'projectName' => $formData['form']['project_name'] ?? null,
+                'projectKey' => $formData['form']['project_key'] ?? null,
+                'description' => $formData['form']['description'] ?? null,
+                'teamName' => $formData['form']['selectedTeamConfig']['team_name'] ?? null,
+                'account' => $formData['form']['account'] ?? null,
+                'newAccount' => $formData['form']['new_account'] ?? false,
+                'newAccountName' => $formData['form']['new_account_name'] ?? null,
+                'newAccountKey' => $formData['form']['new_account_key'] ?? null,
+                'newAccountContact' => $formData['form']['new_account_contact'] ?? null,
+                'newAccountCustomer' => $formData['form']['new_account_customer'] ?? null,
+                'newCustomer' => $formData['form']['new_customer'] ?? false,
+                'newCustomerName' => $formData['form']['new_customer_name'] ?? null,
+                'newCustomerKey' => $formData['form']['new_customer_key'] ?? null,
             ]
         );
     }
@@ -146,8 +173,7 @@ class CreateProjectController extends AbstractController
     /**
      * Create array of all project names and their keys.
      *
-     * @return array
-     *               All projects indexed by key
+     * @return array All projects indexed by key
      */
     private function allProjectsByKey(): array
     {