+++ src/Form/CreateProjectType.php
@@ -15,7 +15,7 @@ use Symfony\Component\Validator\Constraints\Length;
 use Symfony\Component\Validator\Constraints\NotNull;
 use Symfony\Component\Validator\Constraints\Regex;
 
-class CreateProjectForm extends AbstractType
+class CreateProjectType extends AbstractType
 {
     public function __construct(
         private readonly JiraApiService $apiService,
@@ -27,36 +27,33 @@ class CreateProjectForm extends AbstractType
      *
      * @param FormBuilderInterface $builder
      *                                      The form builder
-     * @param array                $options
+     * @param array $options
      *                                      Options related to the form
      *
      * @throws ApiServiceException
      */
     public function buildForm(FormBuilderInterface $builder, array $options): void
     {
-        $builder->add('project_name', TextType::class, [
-            'label' => 'create_project_form.project_name.label',
-            'label_attr' => [
-                'class' => 'form-label',
-            ],
-            'constraints' => [
-                new NotNull(['groups' => 'base']),
-            ],
-            'attr' => [
-                'class' => 'form-element',
-            ],
-            'help_attr' => [
-                'class' => 'form-help',
-            ],
-            'help' => 'create_project_form.project_name.help',
-            'required' => false,
-            'row_attr' => ['class' => 'form-element-wrapper'],
-        ])
+        $builder
+            ->add('project_name', TextType::class, [
+                'label' => 'create_project_form.project_name.label',
+                'label_attr' => ['class' => 'label'],
+                'constraints' => [
+                    new NotNull(['groups' => 'base']),
+                ],
+                'attr' => [
+                    'class' => 'form-element',
+                ],
+                'help_attr' => [
+                    'class' => 'form-help',
+                ],
+                'help' => 'create_project_form.project_name.help',
+                'required' => true,
+                'row_attr' => ['class' => 'form-element-wrapper'],
+            ])
             ->add('project_key', TextType::class, [
                 'label' => 'create_project_form.project_key.label',
-                'label_attr' => [
-                    'class' => 'form-label',
-                ],
+                'label_attr' => ['class' => 'label'],
                 'constraints' => [
                     new NotNull(['groups' => 'base']),
                     new Regex([
@@ -77,14 +74,12 @@ class CreateProjectForm extends AbstractType
                     'class' => 'form-help',
                 ],
                 'help' => 'create_project_form.project_key.help',
-                'required' => false,
+                'required' => true,
                 'row_attr' => ['class' => 'form-element-wrapper'],
             ])
             ->add('description', TextareaType::class, [
                 'label' => 'create_project_form.description.label',
-                'label_attr' => [
-                    'class' => 'form-label',
-                ],
+                'label_attr' => ['class' => 'label'],
                 'constraints' => [
                     new NotNull(['groups' => 'base']),
                 ],
@@ -101,9 +96,7 @@ class CreateProjectForm extends AbstractType
             ])
             ->add('team', ChoiceType::class, [
                 'label' => 'create_project_form.team.label',
-                'label_attr' => [
-                    'class' => 'form-label',
-                ],
+                'label_attr' => ['class' => 'label'],
                 'choices' => $this->getTeamChoices(),
                 'choice_translation_domain' => false,
                 'constraints' => [
@@ -114,21 +107,24 @@ class CreateProjectForm extends AbstractType
                 ],
                 'attr' => [
                     'class' => 'form-element',
+                    'data-create-project-target' => 'choice',
                 ],
                 'help_attr' => [
                     'class' => 'form-help',
                 ],
                 'help' => 'create_project_form.team.help',
+                'required' => true,
                 'row_attr' => ['class' => 'form-element-wrapper'],
             ])
             ->add('account', ChoiceType::class, [
                 'label' => 'create_project_form.account.label',
-                'label_attr' => [
-                    'class' => 'form-label',
-                ],
+                'label_attr' => ['class' => 'label'],
                 'choices' => $this->getAccountChoices(),
                 'choice_translation_domain' => false,
-                'attr' => ['class' => 'form-element js-select2'],
+                'attr' => [
+                    'class' => 'form-element',
+                    'data-create-project-target' => 'choice',
+                ],
                 'help_attr' => [
                     'class' => 'form-help',
                 ],
@@ -139,15 +135,15 @@ class CreateProjectForm extends AbstractType
                         'message' => 'create_project_form.account.constraint.not_null',
                     ]),
                 ],
+                'required' => true,
                 'row_attr' => ['class' => 'form-element-wrapper'],
             ])
             ->add('new_account', CheckboxType::class, [
                 'label' => 'create_project_form.create_new_account.label',
-                'label_attr' => ['class' => 'form-label'],
+                'label_attr' => ['class' => 'label'],
                 'attr' => [
                     'class' => 'form-check-input',
-                    'data-toggle' => 'collapse',
-                    'data-target' => '.toggle-account-group',
+                    'data-action' => 'change->create-project#toggleNewAccount',
                 ],
                 'help_attr' => [
                     'class' => 'form-help',
@@ -159,9 +155,7 @@ class CreateProjectForm extends AbstractType
             ])
             ->add('new_account_name', TextType::class, [
                 'label' => 'create_project_form.new_account_name.label',
-                'label_attr' => [
-                    'class' => 'form-label',
-                ],
+                'label_attr' => ['class' => 'label'],
                 'attr' => [
                     'class' => 'form-element',
                 ],
@@ -173,15 +167,16 @@ class CreateProjectForm extends AbstractType
                 'constraints' => [
                     new NotNull(['groups' => 'account']),
                 ],
-                'row_attr' => ['class' => 'form-element-wrapper form-group-account hidden'],
+                'row_attr' => [
+                    'class' => 'form-element-wrapper hidden',
+                    'data-create-project-target' => 'account',
+                ],
             ])
             ->add('new_account_key', TextType::class, [
                 'label' => 'create_project_form.new_account_key.label',
-                'label_attr' => [
-                    'class' => 'form-label',
-                ],
+                'label_attr' => ['class' => 'label'],
                 'attr' => [
-                    'class' => 'form-element form-group-account',
+                    'class' => 'form-element',
                 ],
                 'required' => false,
                 'help_attr' => [
@@ -191,35 +186,40 @@ class CreateProjectForm extends AbstractType
                 'constraints' => [
                     new NotNull(['groups' => 'account']),
                 ],
-                'row_attr' => ['class' => 'form-element-wrapper form-group-account hidden'],
+                'row_attr' => [
+                    'class' => 'form-element-wrapper hidden',
+                    'data-create-project-target' => 'account',
+                ],
             ])
             ->add('new_account_contact', TextType::class, [
                 'label' => 'create_project_form.new_account_contact.label',
-                'label_attr' => [
-                    'class' => 'form-label',
-                ],
+                'label_attr' => ['class' => 'label'],
                 'attr' => [
                     'class' => 'form-element',
                 ],
                 'required' => false,
                 'help_attr' => [
-                    'class' => 'form-help form-group-account',
+                    'class' => 'form-help',
                 ],
                 'help' => 'create_project_form.new_account_contact.help',
                 'constraints' => [
                     new NotNull(['groups' => 'account']),
                 ],
-                'row_attr' => ['class' => 'form-element-wrapper form-group-account hidden'],
+                'row_attr' => [
+                    'class' => 'form-element-wrapper hidden',
+                    'data-create-project-target' => 'account',
+                ],
             ])
             ->add('new_account_customer', ChoiceType::class, [
                 'label' => 'create_project_form.new_account_customer.label',
-                'label_attr' => [
-                    'class' => 'form-label',
-                ],
+                'label_attr' => ['class' => 'label'],
                 'choices' => $this->getCustomerChoices(),
                 'choice_translation_domain' => false,
                 'required' => false,
-                'attr' => ['class' => 'form-element js-select2'],
+                'attr' => [
+                    'class' => 'form-element',
+                    'data-create-project-target' => 'choice',
+                ],
                 'help_attr' => [
                     'class' => 'form-help',
                 ],
@@ -230,17 +230,18 @@ class CreateProjectForm extends AbstractType
                         'message' => 'create_project_form.new_account_customer.constraint.not_null',
                     ]),
                 ],
-                'row_attr' => ['class' => 'form-element-wrapper form-group-account hidden'],
+                'row_attr' => [
+                    'class' => 'form-element-wrapper hidden',
+                    'data-create-project-target' => 'account',
+                ],
             ])
             ->add('new_customer', CheckboxType::class, [
                 'label' => 'create_project_form.new_customer.label',
-                'label_attr' => [
-                    'class' => 'form-label',
-                ],
+                'label_attr' => ['class' => 'label'],
                 'attr' => [
                     'class' => 'form-check-input',
-                    'data-toggle' => 'collapse',
-                    'data-target' => '.toggle-customer-group',
+                    'data-create-project-target' => 'customerCheckbox',
+                    'data-action' => 'change->create-project#toggleNewCustomer',
                 ],
                 'required' => false,
                 'help_attr' => [
@@ -248,13 +249,14 @@ class CreateProjectForm extends AbstractType
                 ],
                 'help' => 'create_project_form.new_customer.help',
                 'validation_groups' => ['customer'],
-                'row_attr' => ['class' => 'form-element-wrapper hidden'],
+                'row_attr' => [
+                    'class' => 'form-element-wrapper hidden',
+                    'data-create-project-target' => 'account',
+                ],
             ])
             ->add('new_customer_name', TextType::class, [
                 'label' => 'create_project_form.new_customer_name.label',
-                'attr' => [
-                    'class' => 'form-element',
-                ],
+                'label_attr' => ['class' => 'label'],
                 'required' => false,
                 'help_attr' => [
                     'class' => 'form-help',
@@ -263,13 +265,14 @@ class CreateProjectForm extends AbstractType
                 'constraints' => [
                     new NotNull(['groups' => 'customer']),
                 ],
-                'row_attr' => ['class' => 'form-element-wrapper form-group-customer hidden'],
+                'row_attr' => [
+                    'class' => 'form-element-wrapper hidden',
+                    'data-create-project-target' => 'customer',
+                ],
             ])
             ->add('new_customer_key', TextType::class, [
                 'label' => 'create_project_form.new_customer_key.label',
-                'label_attr' => [
-                    'class' => 'form-label',
-                ],
+                'label_attr' => ['class' => 'label'],
                 'attr' => [
                     'class' => 'form-element',
                 ],
@@ -286,7 +289,10 @@ class CreateProjectForm extends AbstractType
                         'groups' => 'customer',
                     ]),
                 ],
-                'row_attr' => ['class' => 'form-element-wrapper form-group-customer hidden'],
+                'row_attr' => [
+                    'class' => 'form-element-wrapper hidden',
+                    'data-create-project-target' => 'customer',
+                ],
             ]);
 
         $builder->add('save', SubmitType::class, [